Conflicts:
	recipes/alsa/alsa-state.bb
	recipes/angstrom/angstrom-libc-fixup-hack.bb
	recipes/base-files/base-files_3.0.14.bb
	recipes/busybox/busybox.inc
	recipes/busybox/busybox_1.15.3.bb
	recipes/cdrkit/cdrkit_1.1.9.bb
	recipes/eglibc/eglibc_2.11.bb
	recipes/glib-2.0/glib-2.0_2.24.1.bb
	recipes/gnome/libsoup-2.4_2.29.91.bb
	recipes/gstreamer/gst-plugins-bad_0.10.19.bb
	recipes/gstreamer/gst-plugins-bad_0.10.21.bb
	recipes/gstreamer/gst-plugins-base_0.10.29.bb
	recipes/gstreamer/gst-plugins-base_0.10.32.bb
	recipes/gstreamer/gst-plugins-good_0.10.23.bb
	recipes/gstreamer/gst-plugins-ugly_0.10.15.bb
	recipes/gstreamer/gst-plugins.inc
	recipes/gstreamer/gstreamer_0.10.32.bb
	recipes/initscripts/initscripts_1.0.bb
	recipes/klibc/klcc-cross_1.5.20.bb
	recipes/libsigc++-1.2/libsigc++-1.2_1.2.5.bb
	recipes/libsigc++-1.2/libsigc++-1.2_1.2.7.bb
	recipes/libxmlccwrap/libxmlccwrap_0.0.12.bb
	recipes/madwifi/madwifi-ng_r.inc
	recipes/madwifi/madwifi-ng_r3314-20080131.bb
	recipes/mjpegtools/files/mjpegtools-v4l-doesnt-mean-x11.patch
	recipes/mjpegtools/mjpegtools_1.9.0.bb
	recipes/ncurses/ncurses_5.7.bb
	recipes/netkit-base/netkit-base-0.17/gcc4_buildfix.patch
	recipes/netkit-base/netkit-base_0.17.bb
	recipes/openssl/openssl.inc
	recipes/opkg/opkg.inc
	recipes/python/python-2.6-manifest.inc
	recipes/python/python-pyopenssl_0.8.bb
	recipes/python/python-twisted_8.2.0.bb
	recipes/python/python-wifi_0.5.0.bb
	recipes/python/python_2.6.5.bb
	recipes/samba/samba_3.2.15.bb
	recipes/samba/samba_3.5.6.bb
	recipes/streamripper/streamripper_1.64.6.bb
	recipes/sysvinit/sysvinit_2.86.bb
	recipes/tasks/task-base.bb
	recipes/tasks/task-boot.bb
	recipes/tcltk/tcl_8.5.8.bb
	recipes/transmission/transmission.inc
	recipes/util-linux-ng/util-linux-ng.inc
	recipes/xbmc/libcdio_0.81.bb
	recipes/zope/python-zopeinterface_3.5.1.bb

+++ lib/oe/unpack.py
@@ -9,7 +9,7 @@ class UnpackError(Exception):
         self.output = output
 
     def __str__(self):
-        return "Unable to unpack '%s' to '%s' (cmd: %s): %s" % \
+        return "Unable to unpack '%s' to '%s' (cmd: %s):\n%s" % \
                (self.filename, self.destdir, self.command, self.output)
 
 def to_boolean(string, default=None):
@@ -47,8 +47,20 @@ def subprocess_setup():
     # non-Python subprocesses expect.
     signal.signal(signal.SIGPIPE, signal.SIG_DFL)
 
-def unpack_file(file, destdir, dos=False, env=None):
+def unpack_file(file, destdir, parameters, env=None):
     import subprocess, shutil
+
+    try:
+        dos = to_boolean(parameters.get("dos"), False)
+    except ValueError, exc:
+        bb.fatal("Invalid value for 'dos' parameter for %s: %s" %
+                 (filename, parameters.get("dos")))
+
+    try:
+        unpack = to_boolean(parameters.get("unpack"), True)
+    except ValueError, exc:
+        bb.fatal("Invalid value for 'unpack' parameter for %s: %s" %
+                 (filename, parameters.get("unpack")))
 
     dest = os.path.join(destdir, os.path.basename(file))
     if os.path.exists(dest):
@@ -56,40 +68,47 @@ def unpack_file(file, destdir, dos=False, env=None):
             return True
 
     cmd = None
-    if file.endswith('.tar'):
-        cmd = 'tar x --no-same-owner -f %s' % file
-    elif file.endswith('.tgz') or file.endswith('.tar.gz') or file.endswith('.tar.Z'):
-        cmd = 'tar xz --no-same-owner -f %s' % file
-    elif file.endswith('.tbz') or file.endswith('.tbz2') or file.endswith('.tar.bz2'):
-        cmd = 'bzip2 -dc %s | tar x --no-same-owner -f -' % file
-    elif file.endswith('.gz') or file.endswith('.Z') or file.endswith('.z'):
-        root, ext = os.path.splitext(file)
-        cmd = 'gzip -dc %s > %s' % (file, os.path.basename(root))
-    elif file.endswith('.bz2'):
-        root, ext = os.path.splitext(file)
-        cmd = 'bzip2 -dc %s > %s' % (file, os.path.basename(root))
-    elif file.endswith('.tar.xz'):
-        cmd = 'xz -dc %s | tar x --no-same-owner -f -' % file
-    elif file.endswith('.xz'):
-        root, ext = os.path.splitext(file)
-        cmd = 'xz -dc %s > %s' % (file, os.path.basename(root))
-    elif file.endswith('.zip') or file.endswith('.jar'):
-        cmd = 'unzip -q -o'
-        if dos:
-            cmd = '%s -a' % cmd
-        cmd = "%s '%s'" % (cmd, file)
-    elif os.path.isdir(file):
-        shutil.rmtree(dest, True)
-        shutil.copytree(file, dest, True)
-    else:
-        shutil.copy2(file, dest)
+    if unpack:
+        if file.endswith('.tar'):
+            cmd = 'tar x --no-same-owner -f %s' % file
+        elif file.endswith('.tgz') or file.endswith('.tar.gz') or file.endswith('.tar.Z'):
+            cmd = 'tar xz --no-same-owner -f %s' % file
+        elif file.endswith('.tbz') or file.endswith('.tbz2') or file.endswith('.tar.bz2'):
+            cmd = 'bzip2 -dc %s | tar x --no-same-owner -f -' % file
+        elif file.endswith('.gz') or file.endswith('.Z') or file.endswith('.z'):
+            root, ext = os.path.splitext(file)
+            cmd = 'gzip -dc %s > %s' % (file, os.path.basename(root))
+        elif file.endswith('.bz2'):
+            root, ext = os.path.splitext(file)
+            cmd = 'bzip2 -dc %s > %s' % (file, os.path.basename(root))
+        elif file.endswith('.tar.xz'):
+            cmd = 'xz -dc %s | tar x --no-same-owner -f -' % file
+        elif file.endswith('.xz'):
+            root, ext = os.path.splitext(file)
+            cmd = 'xz -dc %s > %s' % (file, os.path.basename(root))
+        elif file.endswith('.tar.lz'):
+            cmd = 'lzip -dc %s | tar x --no-same-owner -f -' % file
+        elif file.endswith('.lz'):
+            root, ext = os.path.splitext(file)
+            cmd = 'lzip -dc %s > %s' % (file, os.path.basename(root))
+        elif file.endswith('.zip') or file.endswith('.jar'):
+            cmd = 'unzip -q -o'
+            if dos:
+                cmd = '%s -a' % cmd
+            cmd = "%s '%s'" % (cmd, file)
+
+    if not unpack or not cmd:
+        if os.path.isdir(file):
+            shutil.rmtree(dest, True)
+            shutil.copytree(file, dest, True)
+        else:
+            shutil.copy2(file, dest)
 
     if not cmd:
         return
 
-    pipe = subprocess.Popen(cmd, preexec_fn=subprocess_setup, shell=True,
-                            cwd=destdir, env=env, stdout=subprocess.PIPE,
-                            stderr=subprocess.STDOUT)
-    stdout = pipe.communicate()[0]
-    if pipe.returncode != 0:
-        raise UnpackError(file, destdir, cmd, stdout)
+    import oe.process
+    try:
+        oe.process.run(cmd, env=env, cwd=destdir)
+    except oe.process.CmdError, exc:
+        raise UnpackError(file, destdir, cmd, str(exc))
